# Improved perf of MemoryExtensions.Contains()