--- a/src/System.Private.CoreLib/shared/System/SpanHelpers.Byte.cs
+++ b/src/System.Private.CoreLib/shared/System/SpanHelpers.Byte.cs
@@ -179,28 +179,6 @@ namespace System
             {
                 if ((int)(byte*)offset < length)
                 {
-                    if ((((nuint)Unsafe.AsPointer(ref searchSpace) + (nuint)offset) & (nuint)(Vector256<byte>.Count - 1)) != 0)
-                    {
-                        // Not currently aligned to Vector256 (is aligned to Vector128); this can cause a problem for searches
-                        // with no upper bound e.g. String.strlen.
-                        // Start with a check on Vector128 to align to Vector256, before moving to processing Vector256.
-                        // This ensures we do not fault across memory pages while searching for an end of string.
-                        Vector128<byte> values = Vector128.Create(value);
-                        Vector128<byte> search = LoadVector128(ref searchSpace, offset);
-
-                        // Same method as below
-                        int matches = Sse2.MoveMask(Sse2.CompareEqual(values, search));
-                        if (matches == 0)
-                        {
-                            // Zero flags set so no matches
-                            offset += Vector128<byte>.Count;
-                        }
-                        else
-                        {
-                            return true;
-                        }
-                    }
-
                     lengthToExamine = GetByteVector256SpanLength(offset, length);
                     if ((byte*)lengthToExamine > (byte*)offset)
                     {