Improved performance of `MemoryExtensions.Contains()` for `byte`s and
`char`s.

``` ini

BenchmarkDotNet=v0.11.5, OS=Windows 10.0.17134.706 (1803/April2018Update/Redstone4)
Intel Core i7-8700K CPU 3.70GHz (Coffee Lake), 1 CPU, 12 logical and 6 physical cores
Frequency=3609378 Hz, Resolution=277.0560 ns, Timer=TSC
.NET Core SDK=3.0.100-preview5-011281
  [Host]     : .NET Core 3.0.0-preview5-27613-06 (CoreCLR 4.6.27613.71, CoreFX 4.700.19.21214), 64bit RyuJIT
  Job-BAAJAC : .NET Core ? (CoreCLR 4.6.27704.0, CoreFX 4.700.19.21911), 64bit RyuJIT
  Job-CTXWTG : .NET Core ? (CoreCLR 4.6.27704.0, CoreFX 4.700.19.21911), 64bit RyuJIT


```
|                  Type |        Method | Toolchain |      Mean |     Error |    StdDev |    Median |       Min |       Max | Ratio |
|---------------------- |-------------- |---------- |----------:|----------:|----------:|----------:|----------:|----------:|------:|
| ContainsByteBenchmark |  ContainsTrue |    Master |  2.240 us | 0.0068 us | 0.0060 us |  2.240 us |  2.227 us |  2.249 us |  1.00 |
| ContainsByteBenchmark |  ContainsTrue |      Test |  1.286 us | 0.0091 us | 0.0085 us |  1.284 us |  1.270 us |  1.302 us |  0.57 |
|                       |               |           |           |           |           |           |           |           |       |
| ContainsCharBenchmark |  ContainsTrue |    Master |  6.953 us | 0.0315 us | 0.0295 us |  6.941 us |  6.917 us |  7.011 us |  1.00 |
| ContainsCharBenchmark |  ContainsTrue |      Test |  6.527 us | 0.0108 us | 0.0095 us |  6.526 us |  6.513 us |  6.546 us |  0.94 |
|                       |               |           |           |           |           |           |           |           |       |
| ContainsByteBenchmark | ContainsFalse |    Master |  3.189 us | 0.0069 us | 0.0058 us |  3.188 us |  3.183 us |  3.203 us |  1.00 |
| ContainsByteBenchmark | ContainsFalse |      Test |  2.936 us | 0.0075 us | 0.0067 us |  2.935 us |  2.929 us |  2.951 us |  0.92 |
|                       |               |           |           |           |           |           |           |           |       |
| ContainsCharBenchmark | ContainsFalse |    Master | 11.950 us | 0.0178 us | 0.0158 us | 11.953 us | 11.929 us | 11.983 us |  1.00 |
| ContainsCharBenchmark | ContainsFalse |      Test |  9.406 us | 0.0569 us | 0.0475 us |  9.391 us |  9.347 us |  9.487 us |  0.79 |